Created a new index page to host email field
